Create a new function to set a single wall
I think using setStoneFailIfNoLibs to set walls was doing too much
stuff related to setting a stone to a color and updating board
state. Ideally we can just empty the positions we want to be walls and
then set them to be walls.

This would probably break if we call at any time except when the game
board is empty, but we don't intend to use it any other way so it's
fine for our purposes (probably!).